Refactor timeline_pane for scroll / maximize
newTimelinePane( ... ) is now responsible for setting up card layout
pane which switches between scroll and maximized layouts. Placing
the card layout outside of the scroll layout will hopefully allow
the rows in the maximized layout to be given null vertical sizes and
allowing them to fill the available screen space.